2/6 html 페이지 리액트 구현 완료
