+++ src/pages/attendance/attendancePage/AttendSidemenu2.css
@@ -1,19 +1,29 @@
 .sidemenu2 {
     width: 245px;
-    height: 100%;
+    height: 141%;
     display: flex;
     flex-direction: column;
     color: #606060;
     background-color: #F5F5F5;
+ 
 }
 
-.sidemenu2_area {
+.sidemenu_back {
     width: 89%;
-    height: 97%;
+    height: 94%;
     position: relative;
     left: 14px;
-    top: 10px;
+    top: 30px;
     background-color: white;
+}
+
+
+.sidemenu2_area {
+    
+    position: relative;
+    left: 14px;
+    top: 90px;
+   
 }
 
 .sidemenu2_title {
@@ -31,9 +41,10 @@
 
 .sidemenu2 ul li {
     display: flex;
+    top: 50px;
     justify-content: center;
     align-items: center;
-    margin: 50px 0px;
+    margin: 60px 0px;
     height: 60px;
     width: 180px;
     border-radius: 10px;
@@ -42,18 +53,20 @@
 
 .sidemenu2 ul div {
     font-size: 20px;
-    width: 230px;
+    width: 190px;
     height: 50px;
     text-align: center;
     background-color: #F5F5F5;
     color: #606060;
     line-height: 50px;
+    position: relative;
+    right: 0px;
 }
 
-.sidemenu2 ul li:hover {
+/* .sidemenu2 ul li:hover {
     background-color: #fa9a85;
     width: 100%;
-}
+} */
 
 .sidemenu2 ul li:hover div {
     color: white;
@@ -71,3 +84,18 @@ a{
   .fade-in.active {
     opacity: 1;
   }
+
+  .attendance_selected_box{
+    top: 105px;
+    pointer-events: none;
+    background: #fa9a85;    
+    position: absolute;
+    border-radius: 10px;
+    margin-bottom: 4px;
+    width: 190px;
+    height: 80px;
+    display: flex;
+    flex-direction: row-reverse;
+    align-items: center;
+    transition: top 0.5s ease-in-out, opacity 0.5s ease-in-out;
+}
