+++ src/pages/attendance/attendancePage/approvalPayment.css
@@ -62,7 +62,7 @@ body {
     border:0.5px solid #606060;
 } 
 
-.approval-box{
+.document-approval-box{
     position: absolute;
     left: 80px;
     background-color: #FFBE98;
@@ -114,7 +114,7 @@ body {
 .list-commute-area{
     position: relative;
     top: 190px;
-    left: 50px;
+    left: 0px;
     width: 1200px;
 }
