+++ src/pages/attendance/attendancePage/rejectPayment.css
@@ -76,7 +76,7 @@ body {
 }
 
 
-.reject-box{
+.my-document-reject-box{
     position: absolute;
     left: 160px;
     background-color: #FFBE98;
@@ -111,10 +111,20 @@ body {
     font-size: 16px;
 }
 
+
+.list-commute-detail table {
+    position: relative;
+    top: 190px;
+    font-size: 16px;
+    width: 1200px; 
+    border-collapse: collapse; /* 테이블 셀 간의 경계를 합칩니다. */
+
+}
+
 .list-commute-area{
     position: relative;
     top: 190px;
-    left: 50px;
+    left: 0px;
     width: 1200px;
 }
